Refactored services to a separate service dir
